BUG: Boolean flags expect same parameters as with GenerateCLP
In a command line tool using GenerateCLP to parse its command line,
boolean are not expected to pass a value. If the flag is specified,
the value is set to true. If the flag is not specified, the value
is set to false, independently of the default value specified in
the XML description file. The default value of booleans is only
used when creating the GUI in Slicer.